Text shorten helper (#745)
* added helper for shortening text

* added helper for shortening text

* made review changes

* made review changes

* made review changes
