+++ src/helpers/UtilityHelper.ts
@@ -57,31 +57,7 @@ export const MaskedChannels = {
   '0xcf3224c64FF2d5b18b5e4460b1103FF9037eFAC2': 1
 }
 
-export const isLengthValid = (data: string, len: number): boolean => {
-  return data.length <= len;
-};
 
-export const isValidUrl = (urlString: string): boolean => {
-  var urlPattern = new RegExp(
-    '^((?:https|http):\\/\\/)' + // validate protocol
-      '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' + // validate domain name
-      '((\\d{1,3}\\.){3}\\d{1,3}))' + // validate OR ip (v4) address
-      '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // validate port and path
-      '(\\?[;&a-z\\d%_.~+=-]*)?' + // validate query string
-      '(\\#[-a-z\\d_]*)?$',
-    'i'
-  ); // validate fragment locator
-  return !!urlPattern.test(urlString);
-};
-
-export const isValidAddress = (address: string): boolean => {
-  return utils.isAddress(address);
-}
-
-// Disable the browser notification on Metamask iphone mobile
-export const isUserAgentIphone = (userAgent:any):boolean => {
-  return userAgent.indexOf('iPhone') !== -1;
-};
 
 export const envUtil = {
   prod: 'app.push.org',
@@ -186,6 +162,10 @@ export const getAliasFromChannelDetails = (channelDetails: Object | null | strin
   }
 
   return {address: null, chainId: null};
+};
+
+export const shortenText = (str: string, substringLengthStart: number, substringLengthEnd?: number):string => {
+  return `${str.substring(0, substringLengthStart)}...${str.substring(str.length-(substringLengthEnd??substringLengthStart))}`;
 };
 
 export default UtilityHelper;
